Move MediaTrackerEventGenerator to internal package
