+++ code/edgemedia/src/test/java/com/adobe/marketing/mobile/edge/media/internal/MediaCollectionTrackerTests.java
@@ -22,7 +22,6 @@ import static org.junit.Assert.fail;
 import com.adobe.marketing.mobile.AdobeCallback;
 import com.adobe.marketing.mobile.Event;
 import com.adobe.marketing.mobile.edge.media.Media;
-import com.adobe.marketing.mobile.edge.media.TestMediaTrackerEventGenerator;
 import com.adobe.marketing.mobile.util.CloneFailedException;
 import com.adobe.marketing.mobile.util.DataReader;
 import com.adobe.marketing.mobile.util.EventDataUtils;
@@ -114,8 +113,7 @@ public class MediaCollectionTrackerTests {
 
         Map<String, Object> config = new HashMap<>();
 
-        mediaTrackerAPIEventGenertor =
-                TestMediaTrackerEventGenerator.create(config, "tracker0", true);
+        mediaTrackerAPIEventGenertor = TestMediaTrackerEventGenerator.create("tracker0", true);
 
         hitProcessor = new FakeMediaHitProcessor();
         tracker = new MediaCollectionTracker(hitProcessor, config);
